% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{apc}
\alias{apc}
\title{Factor Model of Balanced Panel Data}
\usage{
apc(X, kmax)
}
\arguments{
\item{X}{a matrix of size T by N.}

\item{kmax}{integer, indicating the maximum number of factors.}
}
\value{
a list of elements:
\item{X}{the original data}
\item{kmax}{the maximum number of factors}
\item{Fhat}{estimated F}
\item{Lamhat}{estimated Lambda}
\item{Chat}{equals Fhat x Lamhat'}
\item{Dhat}{estimated diagonal matrix D, of dim kmax by kmax}
\item{d}{first kmax elements of Dhat}
\item{d0}{diagonal elements of Dhat}
\item{ehat}{equals X - Chat}
}
\description{
\code{apc} estimates the factor model of a given balanced panel data.
}
\references{
Jushan Bai and Serena Ng (2002), \emph{Determining the number of factors in approximate factor models}.
\url{https://doi.org/10.1111/1468-0262.00273}
}
\author{
Yankang (Bennie) Chen <yankang.chen@yale.edu>

Serena Ng <serena.ng@columbia.edu>

Jushan Bai <jushan.bai@columbia.edu>
}
