% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cde_agency}
\alias{cde_agency}
\title{cde_agency}
\format{
A data frame with x rows and y columns:
\describe{
  \item{column1}{Description of column1.}
  \item{column2}{Description of column2.}
  ...
}
}
\source{
Describe the source of your data here, if applicable.
}
\usage{
cde_agency
}
\description{
A brief description of my dataset.
}
\keyword{datasets}
