% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demeanXY.R
\name{demeanXY}
\alias{demeanXY}
\title{Demean Panel Data}
\usage{
demeanXY(X, N, T, N0, T0)
}
\arguments{
\item{X}{detaframe or matrix of the original panel data.}

\item{N}{integer, total number of columns of the panel data.}

\item{T}{integer, total number of rows of the panel data.}

\item{N0}{integer, the number of columns in the panel data with full data availability.}

\item{T0}{integer, the number of rows in the panel data with full data availability.}
}
\value{
a list of elements:
\item{X1}{demeaned data}
\item{FE}{estimated fixed effects matrix}
}
\description{
\code{demeanXY} demeans the panel data.
}
\author{
Yankang (Bennie) Chen <yankang.chen@yale.edu>

Serena Ng <serena.ng@columbia.edu>

Jushan Bai <jushan.bai@columbia.edu>
}
