% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_qd.R
\name{describe_qd}
\alias{describe_qd}
\title{Describe selected variables in the FRED-QD Data Set}
\usage{
describe_qd(varname, name.only = TRUE, verbose = FALSE)
}
\arguments{
\item{varname}{string or a vector strings of the format \code{"X1"} to
\code{"X135"}.}

\item{name.only}{logical. If \code{TRUE}, return a dataframe with variable
names and types of transformation only;
if \code{FALSE}, return a dataframe with more details.}

\item{verbose}{logical, indicating whether or not descriptions should be printed.}
}
\value{
a vector of variable names, or a data frame with detailed descriptions.
}
\description{
\code{describe_qd} provides a description of the selected variables
in the FRED-QD data set.
}
\examples{
library(fbi)
varnames <- describe_qd(c("X32", "X56"), name.only = TRUE, verbose = FALSE)
}
\references{
Michael W. McCracken and Serena Ng (2020), \emph{FRED-QD Updated Appendix}.
\url{https://s3.amazonaws.com/files.fred.stlouisfed.org/fred-md/FRED-QD_appendix.pdf}
}
\author{
Yankang (Bennie) Chen <yankang.chen@columbia.edu>
}
