% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredmd.R
\name{fredmd}
\alias{fredmd}
\title{Loading FRED-MD Data Set}
\usage{
fredmd(date_start = NULL, date_end = NULL, transform = TRUE,
  local = FALSE)
}
\arguments{
\item{date_start}{Date or \code{NULL}, the start date (included) of the data selection.
If \code{NULL}, select till the latest data available.}

\item{date_end}{Date or \code{NULL}, the end date (included) of the data selection.
If \code{NULL}, select up to the earliest data available.}

\item{transform}{logical, indicating Whether or not the FRED-MD data set
should be transformed according to the transformation code.}

\item{local}{logical, indicating Whether or not the FRED-MD data set
should be loaded from the local files or downloaded online}
}
\value{
a subset of the (transformed) FRED-MD data of class \code{fredmd}.
}
\description{
\code{fredmd} loads the official FRED-MD data set and provides
a few tools to manipulate the data set.
}
\examples{
data <- fredmd(date_start = NULL, date_end = NULL, transform = TRUE, local = FALSE)
}
\references{
Michael W. McCracken and Serena Ng (2015), \emph{FRED-MD and FRED-QD: Monthly and Quarterly Databases for Macroeconomic Research}.
\url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}
}
\author{
Yankang (Bennie) Chen <yankang.chen@columbia.edu>
}
