% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredqd.R
\name{fredqd}
\alias{fredqd}
\title{Loading FRED-QD Data Set}
\usage{
fredqd(file = "", date_start = NULL, date_end = NULL, transform = TRUE)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (either a single string or a raw vector).}

\item{date_start}{Date or \code{NULL}, the start date (included) of the data selection.
If \code{NULL}, select till the latest data available.}

\item{date_end}{Date or \code{NULL}, the end date (included) of the data selection.
If \code{NULL}, select up to the earliest data available.}

\item{transform}{logical, indicating Whether or not the FRED-QD data set
should be transformed according to the transformation code.}
}
\value{
a subset of the (transformed) FRED-QD data of class \code{fredqd}.
}
\description{
\code{fredqd} loads the official FRED-QD data set and provides
a few tools to manipulate the data set.
}
\references{
Michael W. McCracken and Serena Ng (2015), \emph{FRED-MD and FRED-QD: Monthly and Quarterly Databases for Macroeconomic Research}.
\url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}
}
\author{
Yankang (Bennie) Chen <yankang.chen@yale.edu>
}
