% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{get_agency_info}
\alias{get_agency_info}
\title{Get information about the selected agency}
\usage{
get_agency_info(agency, state = NULL, ori_only = FALSE,
  exact_match = TRUE)
}
\arguments{
\item{agency}{A string or vector of strings with the name of the agency you want to lookup (capitalized is ignored).}

\item{state}{A string or vector of strings of state names. If used, returns only agencies in that state. This is useful in cases where multiple agencies have the same name in different states and you only want specific states.}

\item{ori_only}{If TRUE (not default), returns only the ORI and the agency_name columns.}

\item{exact_match}{If TRUE (default), finds matches based on exact match of agency name. Else,
uses `grep()` to find agencies with similar names to inputted agency.}
}
\value{
A data.frame with information about the agency - including ORI code and geographic information. The agency will have as many rows as agencies matched from the `agency` input.
}
\description{
Get information about the selected agency including the 9-digit ORI, geographic information, type of agency, and whether they report to NIBRS.
}
\examples{
get_agency_info("Oakland Police Department")
get_agency_info("Oakland Police Department", state = "california")
}
