% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_crime.R
\name{get_estimated_crime}
\alias{get_estimated_crime}
\title{Get state- or national-level estimated crime counts from the UCR's Offenses Known
and Clearances by Arrest data set.}
\usage{
get_estimated_crime(state_abb = NULL, key = get_api_key())
}
\arguments{
\item{state_abb}{String or vector of strings input for state abbreviation(s) to get data for.
If NULL (default) returns national data.}

\item{key}{A string containing your FBI's Crime Data Explorer API key}
}
\value{
A data.frame with state-level estimated UCR crime data for selected state
}
\description{
Get state- or national-level estimated crime counts from the UCR's Offenses Known
and Clearances by Arrest data set.
}
\examples{
\dontrun{
get_estimated_crime("CA")
}
}
\seealso{
Other UCR crime functions: 
\code{\link{get_agency_crime}()},
\code{\link{get_estimated_arson}()}
}
\concept{UCR crime functions}
