% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{get_nibrs_offense}
\alias{get_nibrs_offense}
\title{Title}
\usage{
get_nibrs_offense(api_key, offense = "robbery", ori = NULL,
  region_name = NULL, state_abb = NULL, nation = TRUE)
}
\arguments{
\item{nation}{}
}
\description{
Title
}
