% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{get_nibrs_victim_offences}
\alias{get_nibrs_victim_offences}
\title{Title}
\usage{
get_nibrs_victim_offences(
  ori = "CA0160000",
  from = 1985,
  to = 2022,
  key = get_api_key()
)
}
\arguments{
\item{ori}{desc}

\item{from}{desc}

\item{to}{desc}

\item{key}{desc}
}
\value{
desc
}
\description{
Title
}
\examples{
desc
}
