% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_police_employment.R
\name{get_police_employment}
\alias{get_police_employment}
\title{Get agency-, state-, region-, or national-level police staffing data.}
\usage{
get_police_employment(key, ori = NULL, state_abb = NULL,
  region = NULL)
}
\arguments{
\item{key}{Crime Data Explorer API key}

\item{state_abb}{String or vector of strings input for state abbreviation(s) to get data for -
if NULL returns national data. By default will get all state's data.}

\item{region}{}
}
\value{
A data.frame with columns for annual number of employees and officers (also broken up by gender).
}
\description{
Get agency-, state-, region-, or national-level police staffing data.
}
