% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{is_valid_ori}
\alias{is_valid_ori}
\title{Checks if the ORI is valid}
\usage{
is_valid_ori(ori)
}
\arguments{
\item{ori}{A string or vector of strings of the 9-character ORI code (unique agency ID) for the desired agency.}
}
\value{
TRUE if ori is valid, else FALSE. If ori is a vector, returns TRUE or FALSE
for each ori.
}
\description{
Checks if ORI (a 9-digit unique identifier code assigned to each agency)
is valid for the API.
}
\examples{
# ORI for Oakland Police Deparment in California
is_valid_ori("CA0010900")
# Incorrect ori
is_valid_ori("abc123")
}
