% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{list_nibrs_offense_variables}
\alias{list_nibrs_offense_variables}
\title{Returns a vector of all `variable` parameter options for `get_nibrs_offense()`.}
\usage{
list_nibrs_offense_variables()
}
\value{
A vector of all `variable` parameter options for `get_nibrs_offense()`.
}
\description{
Returns a vector of all `variable` parameter options for `get_nibrs_offense()`.
}
\examples{
list_nibrs_offender_variables()
}
