% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{list_nibrs_offenses}
\alias{list_nibrs_offenses}
\title{Return a vector of all offenses in the NIBRS data.}
\usage{
list_nibrs_offenses()
}
\value{
A vector containing all permissible offenses to input to `get_nibrs_offenses()`,
`get_nibrs_victim()`, and `get_nibrs_offender()`.`
}
\description{
Return a vector of all offenses in the NIBRS data.
}
\examples{
list_nibrs_offenses()
}
