% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{list_nibrs_victim_variables}
\alias{list_nibrs_victim_variables}
\title{Returns a vector of all `variable` parameter options for `get_nibrs_victim()`.}
\usage{
list_nibrs_victim_variables()
}
\value{
A vector of all `variable` parameter options for `get_nibrs_victim()`.
}
\description{
Returns a vector of all `variable` parameter options for `get_nibrs_victim()`.
}
\examples{
list_nibrs_victim_variables()
}
