% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{list_regions}
\alias{list_regions}
\title{Return a vector of all regions choices.}
\usage{
list_regions()
}
\value{
A vector of all regions choices.
}
\description{
Return a vector of all regions choices.
}
\examples{
list_regions()
}
