% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{list_ucr_arrest_offenses}
\alias{list_ucr_arrest_offenses}
\title{Return a vector of all offenses in the UCR Arrest data.}
\usage{
list_ucr_arrest_offenses()
}
\value{
A vector containing all permissible offenses to input to `get_agency_arrest()`.
}
\description{
Return a vector of all offenses in the UCR Arrest data.
}
\examples{
list_ucr_arrest_offenses()
}
