% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nibrs_offender_variables}
\alias{nibrs_offender_variables}
\title{All available variables for NIBRS offender data.}
\format{
A vector with 5 elements:
}
\source{
\url{https://crime-data-explorer.fr.cloud.gov/api}
}
\usage{
nibrs_offender_variables
}
\description{
The `get_nibrs_offender()` function requires a string input for
which variable you want data on. This is a vector of strings
with all possible variables.
}
\keyword{datasets}
