% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Regions available to get estimated UCR and NIBRS data for.}
\format{A vector with 6 elements:}
\source{
\url{https://crime-data-explorer.fr.cloud.gov/api}
}
\usage{
regions
}
\description{
Some functions let you get regional estimated data and require a
string input for which region you want data from. This is a vector
of strings for the six regions available.
}
\keyword{datasets}
