% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_overlay.R
\name{res_overlay.twtp}
\alias{res_overlay.twtp}
\title{Residual Overlay}
\usage{
res_overlay.twtp(object, method = 1, S = 500)
}
\arguments{
\item{object}{an object of class `tw` or `tp`, i.e. the output of
\code{\link{tw_apc}} or \code{\link{tp_apc}}.}

\item{method}{integer 1 to 4, indicating which residual overlay
method to use. They correspond to the four methods described in the paper.}

\item{S}{the number of iterations.}
}
\value{
a list of elements:
\item{method}{the method of residual overlay}
\item{S}{the number of iterations}
\item{cov}{estimated covariance matrix}
\item{cor}{estimated correlation matrix}
}
\description{
\code{res_overlay.twtp} estimates the covariance and correlation matrix of the
unbalanced panel data using the method of residual overlay.
}
\references{
Cahan, E., Bai, J. and Ng, S. 2019, Factor Based Imputation of Missing Data
and Covariance Matrix Estimation. unpublished manuscript, Columbia University
}
\author{
Yankang (Bennie) Chen <yankang.chen@yale.edu>

Serena Ng <serena.ng@columbia.edu>

Jushan Bai <jushan.bai@columbia.edu>
}
