% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_outliers.R
\name{rm_outliers.fredmd}
\alias{rm_outliers.fredmd}
\title{Remove outliers of the FRED-MD Data Set}
\usage{
rm_outliers.fredmd(object)
}
\arguments{
\item{object}{an object of class \code{\link{fredmd}}.}
}
\value{
FRED-MD data of class \code{fredmd} with outliers removed.
}
\description{
\code{rm_outliers.fredmd} removes outliers of the FRED-MD data set produced by
the \code{\link{fredmd}} function.
}
\examples{
data <- fredmd(date_start = NULL, date_end = NULL, transform = TRUE)
newdata <- rm_outlier.fredmd(data)
}
\references{
Michael W. McCracken and Serena Ng (2015), \emph{FRED-MD and FRED-QD: Monthly and Quarterly Databases for Macroeconomic Research}.
\url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}
}
\author{
Yankang (Bennie) Chen <yankang.chen@columbia.edu>
}
