% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.tw.R
\name{se.tw}
\alias{se.tw}
\title{Standard Error of Selected Points (TW)}
\usage{
se.tw(object, npoints, tpoints, qq, re_estimate)
}
\arguments{
\item{object}{an object of class `tw`.}

\item{npoints}{integer or vector of integers, indicating i of the
(i,t) pair(s) of interest.}

\item{tpoints}{integer or vector of integers, indicating t of the
(i,t) pair(s) of interest.}

\item{qq}{placeholder.}

\item{re_estimate}{logical. If `FALSE`, use first pass estimation. If `TRUE`,
use re-estimation.}
}
\value{
a list of elements:
\item{tpoints}{t's of the (i,t) pair(s) of interest}
\item{npoints}{i's of the (i,t) pair(s) of interest}
\item{re_estimate}{logical. If `FALSE`, use first pass estimation; if `TRUE`,
use re-estimation.}
\item{Fhat}{estimated F}
\item{Lamhat}{estimated Lambda}
\item{Chat}{euqals Fhat x Lamhat'}
\item{SigmaC}{estimated variance of C}
\item{SigmaF}{estimated variance of F}
\item{SigmaL}{estimated variance of L}
}
\description{
\code{se.tw} produces the estimated standard error of C^hat
produced by the \code{\link{tw_apc}} function.
}
\references{
Jushan Bai and Serena Ng (2002), \emph{Determining the number of factors in approximate factor models}.
\url{https://doi.org/10.1111/1468-0262.00273}

Jushan Bai and Serena Ng (2019), \emph{Rank regularized estimation of approximate factor models}.
\url{https://doi.org/10.1016/j.jeconom.2019.04.021}

Jushan Bai and Serena Ng (2021), \emph{Matrix Completion, Counterfactuals, and Factor Analysis of Missing Data}.
\url{https://www.tandfonline.com/doi/full/10.1080/01621459.2021.1967163}

Ercument Cahan, Jushan Bai, and Serena Ng (2021),
\emph{Factor-Based Imputation of Missing Values and Covariances in Panel Data of Large Dimensions}.
\url{https://arxiv.org/abs/2103.03045}
}
\author{
Yankang (Bennie) Chen <yankang.chen@yale.edu>

Serena Ng <serena.ng@columbia.edu>

Jushan Bai <jushan.bai@columbia.edu>
}
