% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnt.R
\name{tnt}
\alias{tnt}
\title{Estimate Treatment Effect}
\usage{
tnt(data, param)
}
\arguments{
\item{data}{list containing x1, x2, y0, y1, N0, N1, T0, and T1.}

\item{param}{list containing K, r, do_FE, do_IFE, and maxit1.}
}
\value{
a list of elements:
\item{est}{}
\item{SE}{}
\item{V}{}
\item{it1}{}
}
\description{
\code{tnt} estimates the treatment effect.
}
\references{
Jushan Bai and Serena Ng (2019), \emph{Matrix Completion, Counterfactuals, and Factor Analysis of Missing Data}.
\url{https://arxiv.org/abs/1910.06677}
}
\author{
Yankang (Bennie) Chen <yankang.chen@columbia.edu>

Serena Ng <serena.ng@columbia.edu>

Jushan Bai <jushan.bai@columbia.edu>
}
