% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\docType{data}
\name{ucr_arrest_offenses}
\alias{ucr_arrest_offenses}
\title{All offenses available to get NIBRS data for.}
\format{A vector with 47 elements:}
\source{
\url{https://crime-data-explorer.fr.cloud.gov/api}
}
\usage{
ucr_arrest_offenses
}
\description{
This is a vector of strings with the names of every offense available
to get arrest data for from the `get_arrest_demographics()` function.
}
\keyword{datasets}
