% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{apc}
\alias{apc}
\title{Factor Model of Balanced Panel Data}
\usage{
apc(X, r)
}
\arguments{
\item{X}{a matrix of size T by N.}

\item{r}{integer, indicating the maximum number of factors.}
}
\value{
a list of elements:
\item{Fhat}{}
\item{Lamhat}{}
\item{d}{}
\item{d0}{}
\item{ehat}{}
}
\description{
\code{apc} estiamtes the factor model of a given balanced panel data.
}
\examples{
results <- apc(X, r)
}
\references{
Jushan Bai and Serena Ng (2019), \emph{Matrix Completion, Counterfactuals, and Factor Analysis of Missing Data}.
\url{https://arxiv.org/abs/1910.06677}
}
\author{
Yankang (Bennie) Chen <yankang.chen@columbia.edu>

Serena Ng <serena.ng@columbia.edu>

Jushan Bai <jushan.bai@columbia.edu>
}
