% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbi.R
\docType{package}
\name{fbi-package}
\alias{fbi-package}
\alias{fbi}
\title{Factor-Based Imputation and FRED-MD/QD Data Set}
\description{
The \code{fbi} package contains functions to estimate factor
models and impute missing data based on factor models. It also includes
functions to load and prepare the FRED-MD/QD data set.
}
\details{
See \code{vignette("factor_fred", package = "fbi")}
for an example using the FRED-MD dataset
(\url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}).
}
\references{
Jushan Bai and Serena Ng (2002), \emph{Determining the number of factors in approximate factor models}.
\url{https://onlinelibrary.wiley.com/doi/pdf/10.1111/1468-0262.00273}

Jushan Bai and Serena Ng (2019), \emph{Rank regularized estimation of approximate factor models}.
\url{https://doi.org/10.1016/j.jeconom.2019.04.021}

Jushan Bai and Serena Ng (2021), \emph{Matrix Completion, Counterfactuals, and Factor Analysis of Missing Data}.
\url{https://www.tandfonline.com/doi/full/10.1080/01621459.2021.1967163}

Ercument Cahan, Jushan Bai, and Serena Ng (2021),
\emph{Factor-Based Imputation of Missing Values and Covariances in Panel Data of Large Dimensions}.
\url{https://arxiv.org/abs/2103.03045}
}
\author{
Yankang (Bennie) Chen <yankang.chen@yale.edu>

Serena Ng <serena.ng@columbia.edu>

Jushan Bai <jushan.bai@columbia.edu>
}
