% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_crime.R
\name{get_estimated_arson}
\alias{get_estimated_arson}
\title{Get state-, region- or national-level estimated arson data from the UCR's Offenses Known
and Clearances by Arrest data set.}
\usage{
get_estimated_arson(key = get_api_key(), state_abb = NULL,
  region = NULL)
}
\arguments{
\item{key}{Crime Data Explorer API key}

\item{state_abb}{String or vector of strings input for state abbreviation(s) to get data for -
if NULL returns national data. By default will get all state's data.}

\item{region}{String input for region - if NULL and `state_abb` is NULL, returns national data.}
}
\description{
Get state-, region- or national-level estimated arson data from the UCR's Offenses Known
and Clearances by Arrest data set.
}
\examples{
get_estimated_arson("CA")
}
\seealso{
Other UCR crime functions: \code{\link{get_agency_crime}},
  \code{\link{get_estimated_crime}}
}
\concept{UCR crime functions}
