% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimated_ucr_crime.R
\name{get_estimated_crime}
\alias{get_estimated_crime}
\title{Title}
\usage{
get_estimated_crime(start_year, end_year, api_key, region_name = NULL,
  state_abb = NULL, nation = TRUE, arson = FALSE)
}
\arguments{
\item{arson}{}
}
\description{
Title
}
