% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_crime.R
\name{get_estimated_crime}
\alias{get_estimated_crime}
\title{Get state- or national-level estimated crime counts from the UCR's Offenses Known
and Clearances by Arrest data set.}
\usage{
get_estimated_crime(state_abb = NULL, key = get_api_key())
}
\arguments{
\item{state_abb}{String or vector of strings input for state abbreviation(s) to get data for -
if NULL returns national data. By default will get all state's data.}

\item{key}{}
}
\description{
Get state- or national-level estimated crime counts from the UCR's Offenses Known
and Clearances by Arrest data set.
}
\examples{
get_estimated_crime("CA")
}
\seealso{
Other UCR crime functions: \code{\link{get_agency_crime}},
  \code{\link{get_estimated_arson}}
}
\concept{UCR crime functions}
