% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{get_nibrs_victim}
\alias{get_nibrs_victim}
\title{Title}
\usage{
get_nibrs_victim(api_key, offense = "robbery", variable = "race",
  ori = NULL, region_name = NULL, state_abb = NULL, nation = TRUE)
}
\arguments{
\item{nation}{}
}
\description{
Title
}
