% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_police_employement.R
\name{get_police_employment}
\alias{get_police_employment}
\title{Title}
\usage{
get_police_employment(start_year, end_year, api_key, ori = NULL,
  region_name = NULL, state_abb = NULL, nation = TRUE)
}
\arguments{
\item{nation}{}
}
\description{
Title
}
