% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{list_nibrs_offense_variables}
\alias{list_nibrs_offense_variables}
\title{Returns a vector of all `variable` parameter options for `get_nibrs_offense()`.}
\usage{
list_nibrs_offense_variables()
}
\value{
A vector of all `variable` parameter options for `get_nibrs_offense()`.
}
\description{
Returns a vector of all `variable` parameter options for `get_nibrs_offense()`.
}
\examples{
list_nibrs_offender_variables()
}
\seealso{
Other NIBRS functions: 
\code{\link{get_nibrs_offender}()},
\code{\link{get_nibrs_offense}()},
\code{\link{get_nibrs_victim}()},
\code{\link{list_nibrs_offender_variables}()},
\code{\link{list_nibrs_offenses}()},
\code{\link{list_nibrs_victim_variables}()},
\code{\link{list_regions}()},
\code{\link{list_ucr_arrest_offenses}()}
}
\concept{NIBRS functions}
