% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{list_ucr_arrest_offenses}
\alias{list_ucr_arrest_offenses}
\title{Return a vector of all offenses in the UCR Arrest data.}
\usage{
list_ucr_arrest_offenses()
}
\value{
A vector containing all permissible offenses to input to `get_agency_arrest()`.
}
\description{
Return a vector of all offenses in the UCR Arrest data.
}
\examples{
list_ucr_arrest_offenses()
}
\seealso{
Other NIBRS functions: \code{\link{get_nibrs_offender}},
  \code{\link{get_nibrs_offense}},
  \code{\link{get_nibrs_victim}},
  \code{\link{list_nibrs_offender_variables}},
  \code{\link{list_nibrs_offense_variables}},
  \code{\link{list_nibrs_offenses}},
  \code{\link{list_nibrs_victim_variables}},
  \code{\link{list_regions}}
}
\concept{NIBRS functions}
