% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\docType{data}
\name{nibrs_offenses}
\alias{nibrs_offenses}
\title{All offenses available to get NIBRS data for.}
\format{A vector with 72 elements:}
\source{
\url{https://crime-data-explorer.fr.cloud.gov/api}
}
\usage{
nibrs_offenses
}
\description{
This is a vector of strings with the names of every offense available
for the NIBRS data functions.
}
\keyword{datasets}
