% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se.rpca}
\alias{se.rpca}
\title{Standard Error of C^hat}
\usage{
se.rpca(object, xpoints, qq)
}
\arguments{
\item{object}{an object of class \code{\link{rpca}}.}

\item{xpoints}{placeholder.}

\item{qq}{placeholder.}
}
\value{
standard error of C^hat
}
\description{
\code{se.rpca} produces the estimated standard error of C^hat
produced by the \code{\link{rpca}} function.
}
\references{
Jushan Bai and Serena Ng (2002), \emph{Determining the number of factors in approximate factor models}.
\url{https://doi.org/10.1111/1468-0262.00273}

Jushan Bai and Serena Ng (2017), \emph{Rank regularized estimation of approximate factor models}.
\url{https://doi.org/10.1016/j.jeconom.2019.04.021}
}
\author{
Yankang (Bennie) Chen <yankang.chen@columbia.edu>

Serena Ng <serena.ng@columbia.edu>

Jushan Bai <jushan.bai@columbia.edu>
}
