% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_apc.R
\name{tp_apc}
\alias{tp_apc}
\title{Tall-Project Imputation of Missing Value in Panel Data}
\usage{
tp_apc(X1, r1, center = FALSE, standardize = FALSE)
}
\arguments{
\item{X1}{a matrix of size T by N.}

\item{r1}{integer, indicating the maximum number of factors.}

\item{center}{logical, indicating Whether or not X1 should be demeaned}

\item{standardize}{logical, indicating Whether or not X1 should be scaled.}
}
\value{
a list of elements:
\item{Fhat}{}
\item{Lamhat}{}
\item{Chat}{}
\item{data}{}
}
\description{
\code{tp_apc} imputates the missing values in a given panel data using the
method of "Tall-Project".
}
\references{
Cahan, E., Bai, J. and Ng, S. 2019, Factor Based Imputation of Missing Data
and Covariance Matrix Estimation. unpublished manuscript, Columbia University
}
\author{
Yankang (Bennie) Chen <yankang.chen@columbia.edu>

Serena Ng <serena.ng@columbia.edu>

Jushan Bai <jushan.bai@columbia.edu>
}
