% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_apc.R
\name{tw_apc}
\alias{tw_apc}
\title{Tall-Wide Imputation of Missing Value in Panel Data}
\usage{
tw_apc(X1, r1, center = FALSE, standardize = FALSE, re_estimate = TRUE)
}
\arguments{
\item{X1}{a matrix of size T by N.}

\item{r1}{integer, indicating the maximum number of factors.}

\item{center}{logical, indicating whether or not X1 should be demeaned}

\item{standardize}{logical, indicating whether or not X1 should be scaled.}

\item{re_estimate}{logical, indicating whether or not output factors,
`Fhat`, `Lamhat`, and `Chat`, should be re-estimated from the imputed data.}
}
\value{
a list of elements:
\item{Fhat}{}
\item{Lamhat}{}
\item{Chat}{}
\item{data}{}
}
\description{
\code{tw_apc} imputates the missing values in a given panel data using the
method of "Tall-Wide".
}
\references{
Jushan Bai and Serena Ng (2019), \emph{Matrix Completion, Counterfactuals, and Factor Analysis of Missing Data}.
\url{https://arxiv.org/abs/1910.06677}
}
\author{
Yankang (Bennie) Chen <yankang.chen@columbia.edu>

Serena Ng <serena.ng@columbia.edu>

Jushan Bai <jushan.bai@columbia.edu>
}
